refactor(phlo-api): extract trino SQL helpers
